Add types and cleanup
