+++ src/cypherLanguageSupport/language/cypherBaseFunctions.ts
